formatting and some code
added some more formatting and then copied over some code while commenting on what to do with the code.  There are parts of the code that is going to be modified and when adding to my project.  With this implementation it seems like I don't need an extra activity screen or list if I filter for the exact device.